Revert back initial scripts
